Show numbered accounts on subrouter dashboard
